Modifying font awesome and academics color struc
I’ve changed the structure of how the colors for the font awesome and [ademicons](https://jpswalsh.github.io/academicons/) load.

Icons now have colors in under the `.social-icons` and  `.post__content` selectors. I’ve moved all the font awesome selectors under `.social-icons` to a new `_icons.scss` file and I import it under those selectors. This way the icons have always a consistent color.

The variables for the colors are in `_variables.scss`.

This way I can use them in post and pages easier and with a consistent color.