Fix figure caption
Now it's always under the figure no matter what.

This is a really weird error since the `<figure>` tag in the include has
to be indented so it doesn't get duplicated when it's under a list.

I also added a `<span></span>` tags to the `figcaption` and the `figure`
are always together.